+++ docs/source/history.rst
@@ -5,6 +5,13 @@
 Beta releases (series 0.X)
 ==========================
 
+2017-10-14. Release 0.17.0
+--------------------------
+
+- Added `fields.TimezoneSelection`:func: as simple way to avoid creating the
+  same `fields.Selection` for timezones all over the place.
+
+
 2017-10-14. Release 0.16.2
 --------------------------